refactor: 💡 move each signing service into its own file
