+++ src/signing/dto/signer-details.dto.ts
@@ -1,13 +1,8 @@
 /* istanbul ignore file */
 
-import { ApiProperty } from '@nestjs/swagger';
 import { IsString } from 'class-validator';
 
 export class SignerDetailsDto {
-  @ApiProperty({
-    description:
-      'The value to be used in the `signer` field. The format depends on what signer is being used',
-  })
   @IsString()
   readonly signer: string;
 }
