+++ src/signing/signing.mock.ts
@@ -3,10 +3,10 @@
 import { FireblocksSigningManager } from '@polymeshassociation/fireblocks-signing-manager';
 import { HashicorpVaultSigningManager } from '@polymeshassociation/hashicorp-vault-signing-manager';
 
-import { SigningService } from '~/signing/signing.service';
+import { SigningService } from '~/signing/services';
 import { MockSigningService } from '~/test-utils/service-mocks';
 
-/*
+/**
  * provides a mock HashicorpVaultSigningManager for testing
  */
 export class MockHashicorpVaultSigningManager {
